Move API to appropriate fragments
