+++ cypress/support/fragments/check-in-actions/checkInActions.js
@@ -1,7 +1,8 @@
+import uuid from 'uuid';
 import { Button, Pane, including, TextField, MultiColumnListCell } from '../../../../interactors';
 import NewInctanceHoldingsItem from '../inventory/newInctanceHoldingsItem';
 import NewUser from '../user/newUser';
-import moment from "moment";
+import { REQUEST_METHOD } from '../../constants';
 
 const loadDetailsButton = Button('Loan details');
 const patronDetailsButton = Button('Patron details');
@@ -42,11 +43,14 @@ export default {
     cy.expect(Pane({ title:  'New fee/fine' }).exists());
     this.returnCheckIn();
   },
-  createItemCheckinApi(itemBarcode, servicePointId) {
-    return cy.createItemCheckinApi({
-      itemBarcode,
-      servicePointId,
-      checkInDate: moment.utc().format(),
+  createItemCheckinApi(body) {
+    return cy.okapiRequest({
+      method: REQUEST_METHOD.POST,
+      path: 'circulation/check-in-by-barcode',
+      body: {
+        id: uuid(),
+        ...body,
+      },
     });
   },
 };
